Backend/requirements.txt
# Core ML and data
torch>=2.0.0,<3.0.0
pandas>=1.5
numpy>=1.24
scikit-learn>=1.2
joblib>=1.2

# Server / API
fastapi>=0.95
uvicorn>=0.22
pydantic>=1.10

# Utilities
PyYAML>=6.0
requests>=2.28
psutil>=5.9
pytest>=7.0

# Optional (ONNX export / runtime)
onnx>=1.15
onnxruntime>=1.18

# Optional profiling
ydata-profiling>=4.2
pandas-profiling>=3.2

# Gradient boosting
lightgbm>=4.0
